+++ src/heavyedge_classify/model.py
@@ -6,6 +6,12 @@ from sklearn.linear_model import RidgeClassifierCV
 from sklearn.model_selection import StratifiedKFold
 from sklearn.pipeline import Pipeline
 
+from .calibration import (
+    IsotonicOvOCalibratedClassifierCV,
+    SigmoidOvOCalibratedClassifierCV,
+    TemperatureCalibratedClassifierCV,
+)
+
 __all__ = [
     "minirocket_classifier",
 ]
@@ -20,9 +26,8 @@ def minirocket_classifier(
     ----------
     n_splits : int
         Number of splits for cross-validation.
-    calibration : str, default="sigmoid"
+    calibration : {"sigmoid", "isotonic", "temperature", "sigmoid_ovo", "isotonic_ovo"}
         Calibration method for the classifier.
-        See :class:`sklearn.calibration.CalibratedClassifierCV` for available methods.
     n_jobs : int, default=None
         Number of jobs to run in parallel.
     verbose : bool, default=False
@@ -32,7 +37,7 @@ def minirocket_classifier(
 
     Returns
     -------
-    model : CalibratedClassifierCV
+    model
         MiniRocket-based probabilistic classifier.
 
     Examples
@@ -54,10 +59,32 @@ def minirocket_classifier(
         ],
         verbose=verbose,
     )
-    model = CalibratedClassifierCV(
-        estimator=pipeline,
-        method=calibration,
-        cv=StratifiedKFold(n_splits=n_splits, shuffle=True, random_state=random_state),
-        n_jobs=n_jobs,
-    )
+    cv = StratifiedKFold(n_splits=n_splits, shuffle=True, random_state=random_state)
+    if calibration in ("sigmoid", "isotonic"):
+        model = CalibratedClassifierCV(
+            estimator=pipeline,
+            method=calibration,
+            cv=cv,
+            n_jobs=n_jobs,
+        )
+    elif calibration == "temperature":
+        model = TemperatureCalibratedClassifierCV(
+            estimator=pipeline,
+            cv=cv,
+            n_jobs=n_jobs,
+        )
+    elif calibration == "sigmoid_ovo":
+        model = SigmoidOvOCalibratedClassifierCV(
+            estimator=pipeline,
+            cv=cv,
+            n_jobs=n_jobs,
+        )
+    elif calibration == "isotonic_ovo":
+        model = IsotonicOvOCalibratedClassifierCV(
+            estimator=pipeline,
+            cv=cv,
+            n_jobs=n_jobs,
+        )
+    else:
+        raise ValueError(f"Unsupported calibration method: {calibration}")
     return model